+++ CHANGELOG.md
@@ -11,6 +11,16 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Container image build error logs are now saved to the DB and served via the API.
 
+### Added
+
+- Allow configuring gRPC keepalive
+- output identifier add in metric response object in `compute_plan perf` view.
+- Prevent use of `__` in asset metadata keys
+
+### Removed
+
+- BREAKING: model categories
+
 ## [0.31.0] 2022-09-26
 
 ### Changed